# apps/docs/apis/openapi.yaml
openapi: 3.1.0
info:
  title: AI GNT API
  version: 0.1.0
paths:
...
  /v1/skill-packs/latest.zip:
    get:
      tags:
      - skill-packs
      summary: Latest Skill Pack
      operationId: latest_skill_pack_v1_skill_packs_latest_zip_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
...
  /v1/org/offboarding/request:
    post:
      tags:
      - org-admin
      summary: Request Offboarding
      description: 'Step 1 of 2. Produces the full export (rules with audit history,

        gaps, staleness, calibration/contradiction findings, ROI counters,

        onboarding events, key/connection metadata — never secrets) and mints

        a confirmation token scoped to this org. Calling this again before

        confirming just re-exports and overwrites the previous token (the old

        one stops working) — there is deliberately no limit on how many times

        an admin can re-request before actually confirming.'
      operationId: request_offboarding_v1_org_offboarding_request_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
      - HTTPBearer: []
  /v1/org/offboarding/confirm:
    post:
      tags:
      - org-admin
      summary: Confirm Offboarding
      description: 'Step 2 of 2 — irreversible. Requires the exact token /request
        just

        issued for THIS org (never another org''s — the key is org-scoped, so a

        caller can never hold a valid token for anyone but their own org),

        unexpired, unused. Deletes apps/store''s rules mirror first (the more

        likely-to-fail-partway network call), then every org-scoped Postgres

        row — see gnt.org_offboarding.delete_org_postgres_data for the full,

        authoritative table list.'
      operationId: confirm_offboarding_v1_org_offboarding_confirm_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmOffboardingRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    ConfirmOffboardingRequest:
      properties:
        confirmation_token:
          type: string
          minLength: 1
          title: Confirmation Token
      type: object
      required:
      - confirmation_token
      title: ConfirmOffboardingRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
servers:
- url: https://api.gntai.dev
  description: Production